requirements.txt
ultralytics
opencv-python
torch
torchvision
transformers
pillow
numpy